# RISC-V: optimize vp8_copy_mem with RVV