Test environment: 8c 1804Mhz i5-1140G7

RVV Impl:

```
% CROSS=riscv64-unknown-linux-gnu- configure --target=riscv64-linux-gcc \
	--enable-debug --enable-gprof && make -j
% time qemu-riscv64 -cpu rv64,v=true,zba=true,vlen=128 -L /path/to/sysroot/ \
	./vpxenc --codec=vp8 -w 352 -h 288 -o akiyol.vpx ./akiyo_cif.yuv
Pass 1/1 frame  300/300   314977B    8399b/f  251981b/s   92226 ms (3.25 fps)
user    1m30.108s
% gprof -abp ./vpxenc ./gmon.out | grep vp8_copy_mem
  1.36     53.09     1.04  1025863     0.00     0.00  vp8_copy_mem16x16_rvv
  0.72     59.01     0.55  1641368     0.00     0.00  vp8_copy_mem8x8_rvv
  0.05     65.95     0.04   764377     0.00     0.00  vp8_copy_mem8x4_rvv
```

C Impl:

```
% CROSS=riscv64-unknown-linux-gnu- configure --target=generic-gnu --enable-debug \
	--enable-gprof && make -j
% time qemu-riscv64 -cpu rv64,v=true,zba=true,vlen=128 -L /path/to/sysroot/ \
	./vpxenc --codec=vp8 -w 352 -h 288 -o akiyol.vpx ./akiyo_cif.yuv
Pass 1/1 frame  300/300   314977B    8399b/f  251981b/s   98417 ms (3.05 fps)
user    1m36.146s
% gprof -abp ./vpxenc ./gmon.out | grep vp8_copy_mem
  0.38     63.96     0.31                             vp8_copy_mem8x4_c
  0.04     70.61     0.03   204336     0.00     0.00  vp8_copy_mem16x16_c
```