@Trumeet 2023-08-11
Benchmark on `Intel Xeon Gold 6278C (64) @ 2.600GHz`:

RVV Impl:

```
% time qemu-riscv64-static -cpu rv64,v=true,vlen=128 -L /.../sysroot/ ./vpxenc --codec=vp8 -w 352 -h 288 -o akiyol.vpx ../akiyo_cif.yuv
vector version is not specified, use the default value v1.0
Pass 1/1 frame  300/300   314977B    8399b/f  251981b/s  105408 ms (2.85 fps)

real    1m45.591s
user    1m45.528s
sys     0m0.060s

% gprof -abp ./vpxenc ./gmon.out | grep vp8_copy_mem
  1.13     62.95     1.01  1025863     0.00     0.00  vp8_copy_mem16x16_rvv
  0.61     71.53     0.54  1641368     0.00     0.00  vp8_copy_mem8x8_rvv
  0.02     77.89     0.02   764377     0.00     0.00  vp8_copy_mem8x4_rvv
```

C Impl:

```
% time qemu-riscv64-static -cpu rv64,v=true,vlen=128 -L /.../sysroot/ ./vpxenc --codec=vp8 -w 352 -h 288 -o akiyol.vpx ../akiyo_cif.yuv
vector version is not specified, use the default value v1.0
Pass 1/1 frame  300/300   314977B    8399b/f  251981b/s  169065 ms (1.77 fps)

real    2m49.286s
user    2m49.179s
sys     0m0.096s

% gprof -abp ./vpxenc ./gmon.out | grep vp8_copy_mem
  0.34    115.69     0.49                             vp8_copy_mem8x4_c
  0.08    122.79     0.12   204336     0.00     0.00  vp8_copy_mem16x16_c
```