@Trumeet 2023-08-12
> 问个小问题，为啥 c 里没有 vp8_copy_mem8x8_ ？

@unicornx 

有的呀： `./vp8/common/reconinter.c:35:void vp8_copy_mem8x8_c(unsigned char *src, int src_stride, unsigned char *dst,`