@unicornx 2023-08-14
> > 问个小问题，为啥 c 里没有 vp8_copy_mem8x8_ ？
> 
> @unicornx
> 
> 有的呀： `./vp8/common/reconinter.c:35:void vp8_copy_mem8x8_c(unsigned char *src, int src_stride, unsigned char *dst,`

我可能没说清楚，我的意思是为啥从 gprof 的结果中有 `vp8_copy_mem8x8_rvv`，但没有 `vp8_copy_mem8x8_c`