@Trumeet 2023-08-15
> > > 问个小问题，为啥 c 里没有 vp8_copy_mem8x8_ ？
> > 
> > 
> > @unicornx
> > 有的呀： `./vp8/common/reconinter.c:35:void vp8_copy_mem8x8_c(unsigned char *src, int src_stride, unsigned char *dst,`
> 
> 我可能没说清楚，我的意思是为啥从 gprof 的结果中有 `vp8_copy_mem8x8_rvv`，但没有 `vp8_copy_mem8x8_c`

@unicornx 

是应该是这样的吧，build 的时候 `vp8_copy_mem8x8` 就被 define 成了 `vp8_copy_mem8x8_rvv` ，C 实现不会被调用了。反汇编应该也能看到没有地方调用了 `vp8_copy_mem8x8_c` 。不过其实我本地测试的时候在一些地方还会看到 gprof 结果里有 `_c` 调用，我不是很明白为什么，但正常情况应该是没有的。